Update backup-jellyfin.ps1
Simple Syntax mistake, removed unnecessary $verified check on line 22.

+++ backup-jellyfin.ps1
@@ -19,7 +19,7 @@ If (-not(Get-service -Name "*Jellyfin*"))) {
     $verified = (-not(Get-Process | Where {$_.ProcessName -like "*jellyfin*"} | select $_))
 }
 else {
-    $verified = If (-not(Get-Service -Name "*Jellyfin*").Status -eq "Stopped") { 
+     If (-not(Get-Service -Name "*Jellyfin*").Status -eq "Stopped") { 
         Write-Host "Waiting for Jellyfin Service to stop" -NoNewline
         Stop-Service -Name "*Jellyfin*" -Force
         Do {
